remove all videos is working correctly from header, remove old icon from menu
