+++ app/src/main/java/org/stepic/droid/view/fragments/DownloadsFragment.java
@@ -1,29 +1,21 @@
 package org.stepic.droid.view.fragments;
 
-import android.app.Dialog;
 import android.app.DownloadManager;
 import android.app.ProgressDialog;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.database.Cursor;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
 import android.support.v4.util.Pair;
-import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ProgressBar;
 
-import com.squareup.otto.Bus;
 import com.squareup.otto.Subscribe;
-import com.yandex.metrica.YandexMetrica;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -45,10 +37,6 @@ import org.stepic.droid.model.Lesson;
 import org.stepic.droid.model.Step;
 import org.stepic.droid.model.VideosAndMapToLesson;
 import org.stepic.droid.receivers.DownloadCompleteReceiver;
-import org.stepic.droid.store.CleanManager;
-import org.stepic.droid.store.operations.DatabaseFacade;
-import org.stepic.droid.util.AppConstants;
-import org.stepic.droid.util.DbParseHelper;
 import org.stepic.droid.util.ProgressHelper;
 import org.stepic.droid.util.StepicLogicHelper;
 import org.stepic.droid.view.adapters.DownloadsAdapter;
@@ -60,9 +48,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ThreadPoolExecutor;
-
-import javax.inject.Inject;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
@@ -78,8 +63,6 @@ public class DownloadsFragment extends FragmentBase {
     public static DownloadsFragment newInstance() {
         return new DownloadsFragment();
     }
-
-    public static final String KEY_STRING_IDS = "step_ids";
 
     @Bind(R.id.empty_downloading)
     View mEmptyDownloadView;
@@ -381,102 +364,15 @@ public class DownloadsFragment extends FragmentBase {
         mDownloadAdapter.notifyDataSetChanged();
     }
 
-    @Override
-    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-        super.onCreateOptionsMenu(menu, inflater);
-        menu.clear();
-        if (mEmptyDownloadView.getVisibility() != View.VISIBLE) {
-            inflater.inflate(R.menu.delete_menu, menu);
-        }
-    }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
-            case R.id.action_delete:
-                ClearVideosDialog dialogFragment = new ClearVideosDialog();
-
-                Bundle bundle = new Bundle();
-                long[] stepIds = new long[mCachedVideoList.size()];
-                int i = 0;
-                for (CachedVideo videoItem : mCachedVideoList) {
-                    stepIds[i++] = videoItem.getStepId();
-                }
-                String stringWithIds = DbParseHelper.parseLongArrayToString(stepIds);
-                bundle.putString(KEY_STRING_IDS, stringWithIds);
-                dialogFragment.setArguments(bundle);
-
-                dialogFragment.show(getFragmentManager(), null);
-                return true;
             default:
                 return super.onOptionsItemSelected(item);
         }
     }
 
-
-    public static class ClearVideosDialog extends DialogFragment {
-
-        @Inject
-        DatabaseFacade mDatabaseFacade;
-        @Inject
-        CleanManager mCleanManager;
-        @Inject
-        Bus mBus;
-
-        @Inject
-        ThreadPoolExecutor threadPoolExecutor;
-
-        @NotNull
-        @Override
-        public Dialog onCreateDialog(Bundle savedInstanceState) {
-            MainApplication.component().inject(this);
-            Bundle bundle = getArguments();
-            String stringIds = bundle.getString(KEY_STRING_IDS);
-            final long[] stepIds = DbParseHelper.parseStringToLongArray(stringIds);
-
-            AlertDialog.Builder builder = new AlertDialog.Builder(getActivity(), R.style.MyAlertDialogStyle);
-            builder.setTitle(R.string.title_clear_cache_dialog)
-                    .setMessage(R.string.clear_videos)
-                    .setPositiveButton(R.string.yes, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            YandexMetrica.reportEvent(AppConstants.METRICA_YES_CLEAR_VIDEOS);
-                            mBus.post(new ClearAllDownloadWithoutAnimationEvent(stepIds));
-                            if (stepIds == null) return;
-
-
-                            AsyncTask task = new AsyncTask() {
-
-                                @Override
-                                protected void onPreExecute() {
-                                    super.onPreExecute();
-                                    mBus.post(new StartDeletingLoadEvent());
-
-                                }
-
-                                @Override
-                                protected Object doInBackground(Object[] params) {
-                                    for (long stepId : stepIds) {
-                                        Step step = mDatabaseFacade.getStepById(stepId);
-                                        mCleanManager.removeStep(step);
-                                    }
-                                    return null;
-                                }
-
-                                @Override
-                                protected void onPostExecute(Object o) {
-                                    mBus.post(new FinishDeletingLoadEvent());
-                                }
-                            };
-                            task.executeOnExecutor(threadPoolExecutor);
-                        }
-                    })
-                    .setNegativeButton(R.string.no, null);
-            setCancelable(false);
-
-            return builder.create();
-        }
-    }
 
     @Subscribe
     public void onClearAll(ClearAllDownloadWithoutAnimationEvent e) {
@@ -549,7 +445,6 @@ public class DownloadsFragment extends FragmentBase {
         } else {
             mEmptyDownloadView.setVisibility(View.VISIBLE);
         }
-        getActivity().invalidateOptionsMenu();
     }
 
     @Subscribe
